pythonAdmin/requirements.txt
customtkinter
requests
python-dotenv
Pillow
